Finished up defense Whole play runs now
